add runtime check to backup DB after version update
The first time the app runs after a version update,
if the database schema is determined to also require an update,
then a backup of the DB should be saved before any modification.

Previously, a runtime permission check was performed silently..
and a backup would only occur if permission to write
to external storage had already been granted.

Now, a more complicated control flow allows the app
to request permission from the user..
and if initially denied, to display an alert dialog..
which explains the implications of not saving a backup,
and provides one final chance to grant the permission.

This is done by allowing the DB update listener
the ability to cancel the update..
which gives the listener the opportunity to manage permissions,
and trigger the update again at a later time.
To prevent an infinite loop, when the listener does restart the update,
an additional parameter is passed which prevents the listener
from being given the opportunity to intercept the update a 2nd time.